chore: update components and how to guides * Move demo app into its own component to easily not use demo app * Moved firebase configuration to environment variables
