+++ firebase.json
@@ -1,6 +1,11 @@
 {
   "hosting": {
     "public": "build",
+    "ignore": [
+      "firebase.json",
+      "**/.*",
+      "**/node_modules/**"
+    ],
     "headers": [
       {
         "source": "/service-worker.js",
