+++ src/components/DemoApp/DemoApp.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import DemoApp from './DemoApp'
 import { BrowserRouter } from 'react-router-dom'
 import renderer from 'react-test-renderer'
 
-describe('components > App', () => {
+describe('components > DemoApp', () => {
   it('renders correctly', () => {
-    const tree = renderer.create(<App />)
+    const tree = renderer.create(<DemoApp />)
     expect(tree.toJSON()).toMatchSnapshot()
   })
 })
